+++ src/components/theme/Anontools/Anontools.test.jsx
@@ -5,17 +5,13 @@ import { Provider } from 'react-intl-redux';
 import { MemoryRouter } from 'react-router-dom';
 
 import Anontools from './Anontools';
-import { arrayWIdsToObject } from '@plone/volto/helpers/Utils/Utils';
 
 const mockStore = configureStore();
 
-const actions = { user: [{ id: 'login' }] };
-const actionsById = arrayWIdsToObject(actions);
-
 describe('Anontools', () => {
-  it('renders an anontools component when no use is logged in', () => {
+  it('renders an anontools component when no token is specified', () => {
     const store = mockStore({
-      actions: { actions, actionsById },
+      userSession: { token: null },
       content: { data: { '@id': 'myid' } },
       intl: {
         locale: 'en',
@@ -33,10 +29,9 @@ describe('Anontools', () => {
     expect(json).toMatchSnapshot();
   });
 
-  it('should not render an anontools component when a user is logged in', () => {
-    const actions = { user: [{ id: 'logout' }] };
+  it('should not render an anontools component when a token is specified', () => {
     const store = mockStore({
-      actions: { actions, actionsById: arrayWIdsToObject(actions) },
+      userSession: { token: '1234' },
       content: { data: {} },
       intl: {
         locale: 'en',