+++ src/selectors/userSession/userSession.js
@@ -12,16 +12,7 @@ import jwtDecode from 'jwt-decode';
  * @returns {boolean} `true` if a user is currently authenticated, `false` otherwise.
  */
 export function loggedIn(state) {
-  /* The user may be authenticated by different means, including outside the UI.  Defer
-   * to the response from Plone, sepcifically whether Plone presents an option to log
-   * in. */
-
-  if (
-    typeof state.actions.actionsById.user.login === 'undefined' &&
-    typeof state.actions.actionsById.user.logout === 'undefined'
-  )
-    return false;
-  return !!state.actions.actionsById.user?.logout;
+  return !!state.userSession.token;
 }
 
 /**